+++ src/main/kotlin/net/pwall/json/stream/JSONArrayCoPipeline.kt
@@ -31,7 +31,8 @@ import net.pwall.util.pipeline.AbstractIntObjectCoPipeline
 import net.pwall.util.pipeline.CoAcceptor
 
 /**
- * A (coroutine) pipeline that accepts Unicode code points and emits `JSONValue`s.
+ * A (coroutine) pipeline that accepts a JSON array as a stream of Unicode code points and emits a `JSONValue` for each
+ * item in the array.
  *
  * @constructor
  * @param   valueConsumer   the `JSONValue` consumer
@@ -41,62 +42,15 @@ import net.pwall.util.pipeline.CoAcceptor
 class JSONArrayCoPipeline<R>(valueConsumer: CoAcceptor<JSONValue?, R>) :
         AbstractIntObjectCoPipeline<JSONValue?, R>(valueConsumer) {
 
-    enum class State { INITIAL, FIRST, ENTRY, COMMA, COMPLETE }
-
-    private var state: State = State.INITIAL
-    private var child: JSONCoValueBuilder = JSONCoValueBuilder()
-
-    override val complete: Boolean
-        get() = state == State.COMPLETE
-
-    override suspend fun acceptInt(value: Int) {
-        when (state) {
-            State.INITIAL -> {
-                if (!JSONCoBuilder.isWhitespace(value)) {
-                    if (value.toChar() == '[')
-                        state = State.FIRST
-                    else
-                        throw JSONException("Pipeline must contain array")
-                }
-            }
-            State.FIRST -> {
-                if (!JSONCoBuilder.isWhitespace(value)) {
-                    if (value.toChar() == ']')
-                        state = State.COMPLETE
-                    else {
-                        state = State.ENTRY
-                        child.acceptChar(value)
-                        // always true for first character
-                    }
-                }
-            }
-            State.ENTRY -> {
-                val consumed = child.acceptChar(value)
-                if (child.complete) {
-                    emit(child.result)
-                    state = State.COMMA
-                }
-                if (!consumed) {
-                    state = State.COMMA
-                    expectComma(value)
-                }
-            }
-            State.COMMA -> expectComma(value)
-            State.COMPLETE -> JSONCoBuilder.checkWhitespace(value)
-        }
+    private val arrayProcessor = JSONArrayProcessor {
+        emit(it)
     }
 
-    private fun expectComma(ch: Int) {
-        if (!JSONCoBuilder.isWhitespace(ch)) {
-            state = when (ch.toChar()) {
-                ',' -> {
-                    child = JSONCoValueBuilder()
-                    State.ENTRY
-                }
-                ']' -> State.COMPLETE
-                else -> throw JSONException("Illegal syntax in JSON array")
-            }
-        }
+    override val complete: Boolean
+        get() = arrayProcessor.complete
+
+    override suspend fun acceptInt(value: Int) {
+        arrayProcessor.acceptInt(value)
     }
 
     override fun close() {